chore: split __tests__/lib.spec.ts into two files, one containing the tests and one containing the importable stuff
