+++ src/__tests__/rundown.spec.ts
@@ -9,7 +9,7 @@ import {
 import { Conductor } from '../conductor'
 import { MockTime } from './mockTime'
 import * as _ from 'underscore'
-import { getMockCall } from './lib.spec'
+import { getMockCall } from './lib'
 
 // let nowActual: number = Date.now()
