+++ src/devices/__tests__/abstract.spec.ts
@@ -8,7 +8,7 @@ import { AbstractDevice } from '../abstract'
 import { StatusCode } from '../device'
 import { MockTime } from '../../__tests__/mockTime'
 import { ThreadedClass } from 'threadedclass'
-import { getMockCall } from '../../__tests__/lib.spec'
+import { getMockCall } from '../../__tests__/lib'
 
 describe('Abstract device', () => {
 	let mockTime = new MockTime()
