+++ src/devices/__tests__/casparcg.spec.ts
@@ -11,7 +11,7 @@ import {
 	Direction
 } from '../../types/src'
 import { MockTime } from '../../__tests__/mockTime'
-import { getMockCall } from '../../__tests__/lib.spec'
+import { getMockCall } from '../../__tests__/lib'
 
 // usage logCalls(commandReceiver0)
 function logCalls (fcn) {
