+++ src/devices/__tests__/lawo.spec.ts
@@ -9,7 +9,7 @@ import {
 } from '../../types/src'
 import { MockTime } from '../../__tests__/mockTime'
 import { ThreadedClass } from 'threadedclass'
-import { getMockCall } from '../../__tests__/lib.spec'
+import { getMockCall } from '../../__tests__/lib'
 
 describe('Lawo', () => {
 	let mockTime = new MockTime()
