+++ src/devices/__tests__/osc.spec.ts
@@ -11,7 +11,7 @@ import {
 import { MockTime } from '../../__tests__/mockTime'
 import { literal } from '../device'
 import { ThreadedClass } from 'threadedclass'
-import { getMockCall } from '../../__tests__/lib.spec'
+import { getMockCall } from '../../__tests__/lib'
 
 // let nowActual = Date.now()
 describe('OSC-Message', () => {
