+++ src/devices/__tests__/pharos.spec.ts
@@ -9,7 +9,7 @@ import {
 } from '../../types/src'
 import { MockTime } from '../../__tests__/mockTime'
 import { ThreadedClass } from 'threadedclass'
-import { getMockCall } from '../../__tests__/lib.spec'
+import { getMockCall } from '../../__tests__/lib'
 const WebSocket = require('../../__mocks__/ws')
 
 describe('Pharos', () => {
