+++ src/devices/__tests__/quantel.spec.ts
@@ -9,7 +9,7 @@ import {
 import { MockTime } from '../../__tests__/mockTime'
 import { ThreadedClass } from 'threadedclass'
 import { QuantelDevice, QuantelCommandType } from '../quantel'
-require('../../__tests__/lib.spec')
+import '../../__tests__/lib'
 
 const orgSetTimeout = setTimeout
